Make allocation of funds transparent to public
After the recent $240 million debacle in Oregon, it occurs to me that there seem to be a large number of projects where we find out about enormous contracts that have failed after all of the money is spent.  The sure way to prevent this is to require that whenever someone gets paid, it is public knowledge.  That way the public can monitor the project and cannot be surprised by a sudden write-off of hundreds of millions of dollars.