Add OSI license requriements
Open source licenses are only licenses that comply with the Open Source Definition — in brief, they allow software to be freely used, modified, and shared. To be approved by the Open Source Initiative (also known as the OSI), a license must go through the Open Source Initiative's license review process and as such is the industry and globally recognized standard for identifying open source software.

The plays should explicitly require that software carry an OSI Approved Open Source License to ensure software freedom, meet the expectations of those involved with the open source community and ensure the benefits open source software and licenses afford. There are many organizations which use the label "open source" but develop their own licenses which do not comply with the OSD and limit the software freedom guaranteed by the OSI and OSD.