Add OSI license requriements
The plays should explicitly require that software carry an Open Source Initiative (OSI) Approved Open Source License to ensure software freedom, meet the expectations of those involved with the open source community and ensure the benefits open source software and licenses afford. There are many organizations which use the label "open source" but develop their own licenses (or may not carry any license) and thus do not comply with the OSD resulting in limitations in software freedom as guaranteed by the OSI and OSD.

Open source licenses are only licenses that comply with the Open Source Definition — in brief, they allow software to be freely used, modified, and shared. To be approved by the OSI, a license must go through the Open Source Initiative's license review process and as such is the industry and globally recognized standard for identifying open source software.