Dead Links + Formatting Fixes
Updated several dead links in main document and footnotes, added brackets to footnote [21], made formatting of footnotes [2] and [4] consistent with style of the rest of the footnotes (i.e. with no space between footnote open bracket and preceding char), one space added elsewhere between period and start of next sentence. Note: Footnote [29] doesn't exist in the body of document, but seems like it should reference the Appendix A .png file, so perhaps add it in the image? I didn't see Justification and Approval mentioned anywhere else in the document but the image file, so there were no logical places to insert it into the text.