Update 02.md
Changes under "Key Questions"
Here, the focus should be the user needs, not a basis of accomplishing a task. Applications will be designed that serve a need, people are not currently solving. Just as the internet solved a globalization need that hadn't been previously addressed. Changing the language to allow for a UX process that solves unmet needs, opens the door for more innovative design.