Update 01.md
Added the word "synthesis" in Checklist 4. "Synthesis" represents the process of combining research and understanding to create theories about what the research means. It's not simply a matter of creating a "prioritized list of tasks" based on what the "user is trying to accomplish." Often, the user does not know what she wants. It's our job, as designers, to understand a user's deep needs and desires and this is why synthesis is vital to the UX process.

Changed Checklist 6. to read "Create a prioritized list of tasks of the user goals." To echo the sentiment above, the user may be "trying to accomplish" something that's not solving a deep need. Oftentimes, the user does not understand what he needs most. "If I had asked people what they want, they would have said faster horses." - Henry Ford. Also, eliminated language about "user stories". The term "user stories" is ambiguous and means different things to developers and designers - even varying in meaning amongst designers. For this reason, it should  be eliminated in all UX conversation.